Mise en place de mediatR

+++ TrainingManagement.RestApi/Domain/Ref/DegreeUv.cs
@@ -6,8 +6,8 @@ namespace TrainingManagement.RestApi.Domain.Ref
 {
     public class DegreeUv : BaseEntity
     {
-        public int DegreeId{ get; set; }
-        public int UvId { get; set; }
+        public long DegreeId{ get; set; }
+        public long UvId { get; set; }
         public virtual Degree Degree { get; set; }
         public virtual Uv Uv { get; set; }
